adds support for multiple ACLK query processing threads (#9355)
